requirements.txt
aiohttp
discord-py-slash-command==2.3.1
discord.py==1.7.2
openai